requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph>=0.1.0
langchain-core>=0.2.0
